🎨 クリップボード機能修正
